Arc Layer

Arc Layer

Usage

add_arc_layer(
  rdeck,
  ...,
  id = uuid::UUIDgenerate(),
  name = "ArcLayer",
  group_name = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  wrap_longitude = FALSE,
  position_format = "XY",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_source_position = source_position,
  get_target_position = target_position,
  get_source_color = "#000000ff",
  get_target_color = "#000000ff",
  get_width = 1,
  get_height = 1,
  get_tilt = 0,
  great_circle = FALSE,
  width_units = "pixels",
  width_scale = 1,
  width_min_pixels = 0,
  width_max_pixels = 9007199254740991,
  blending_mode = "normal",
  visibility_toggle = TRUE,
  tooltip = NULL
)

update_arc_layer(
  rdeck,
  ...,
  id,
  name = cur_value(),
  group_name = cur_value(),
  data = cur_value(),
  visible = cur_value(),
  pickable = cur_value(),
  opacity = cur_value(),
  wrap_longitude = cur_value(),
  position_format = cur_value(),
  color_format = cur_value(),
  auto_highlight = cur_value(),
  highlight_color = cur_value(),
  get_source_position = cur_value(),
  get_target_position = cur_value(),
  get_source_color = cur_value(),
  get_target_color = cur_value(),
  get_width = cur_value(),
  get_height = cur_value(),
  get_tilt = cur_value(),
  great_circle = cur_value(),
  width_units = cur_value(),
  width_scale = cur_value(),
  width_min_pixels = cur_value(),
  width_max_pixels = cur_value(),
  blending_mode = cur_value(),
  visibility_toggle = cur_value(),
  tooltip = cur_value()
)

Arguments

rdeck

<rdeck | rdeck_proxy> An rdeck map instance.

...

Additional parameters that will be forwarded to deck.gl javascript without validation nor processing. All dots must be named and will be camelCased when serialised. A warning is raised when dots are used, warning class rdeck_dots_nonempty.

id

<string> The layer's identifier must be unique for among all layers of the same type for a map. Defaults to uuid::UUIDgenerate(), but should be explicitly defined for updatable layers in a shiny application.

name

<string> Identifies the layer on tooltips and legends. It does not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.

group_name

<string> Defines the group that this layer belongs to. Currently only effective on the layer selector, if visibility_toggle = TRUE.

data

<data.frame | sf | string> The layer's data. Data frames and sf objects will contain all columns that are referenced by the layer's accessors. Strings will be interpreted as a URL and data will be retrieved dynamically in the browser.

visible

<boolean> Determines whether the layer is visible or not; also determines whether any legend elements for the layer will be displayed.

pickable

<boolean> Determines if the layer responds to pointer / touch events.

opacity

<number> Determines the layer's opacity.

wrap_longitude

<boolean> Normalises geometry longitudes.

position_format

<"XY" | "XYZ"> Determines whether each coordinate has two (XY) or three (XYZ) elements.

color_format

<"RGB" | "RGBA"> Determines whether the alpha channel of the colours will be ignored by accessors, making all colours opaque.

auto_highlight

<boolean> When TRUE, the current object hovered by the cursor is highlighted by highlight_color.

highlight_color

<accessor | scale | color> When auto_highlight and pickable are enabled, highlight_color determines the colour of the currently highlighted object. If a single colour value is supplied, that colour will be used to highlight all objects in the layer. Per-object highlighting is achieved with a colour scale, or a tidy-eval column of colours.

get_source_position

<accessor> The feature source positions. A <point/multipoint> wk-geometry column with CRS EPSG:4326. Supports tidy-eval.

get_target_position

<accessor> The feature target positions. A <point/multipoint> wk-geometry column with CRS EPSG:4326. Supports tidy-eval.

get_source_color

<accessor | scale | color> The colour of the source end of the arc. Accepts a single colour value, a colour scale, or a tidy-eval column of colours.

get_target_color

<accessor | scale | color> The colour of the target end of the arc. Accepts a single colour value, a colour scale, or a tidy-eval column of colours.

get_width

<accessor | scale | number> The width of each object, in units specified by width_scale. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers.

get_height

<accessor | scale | number> The multiplier of layer of layer height. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers. A value of 0 will make the layer flat.

get_tilt

<accessor | number> Tilts the arcs by the specified number of degrees (between c(-90, 90)). Accepts a single numeric value or a tidy-eval column of numbers.

great_circle

<boolean> If TRUE, create the arc along the shortest path on the earth surface.

width_units

<"pixels" | "common" | "meters"> The units of the line_width.

width_scale

<number> The scaling multiplier for the width of each line.

width_min_pixels

<number> The minimum line width in pixels.

width_max_pixels

<number> The maximum line width in pixels.

blending_mode

<"normal" | "additive" | "subtractive"> Sets the blending mode. Blending modes:

normal: Normal blending doesn't alter colours of overlapping objects.

additive: Additive blending adds colours of overlapping objects. Useful for highlighting dot density on dark maps.

subtractive: Subtractive blending darkens overlapping objects. Useful for highlighting dot density on light maps.

visibility_toggle

<boolean> Whether this layer will appear in the layer selector.

tooltip

<tooltip> Defines the columns (and their order) that will be displayed in the layer tooltip, if pickable == TRUE. Supports tidy-select if a data is a data.frame. Geometry columns are always removed.

See also