H3 Hexagon Layer

H3 Hexagon Layer

add_h3_hexagon_layer(
  rdeck,
  ...,
  id = NULL,
  name = "H3HexagonLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  stroked = TRUE,
  filled = TRUE,
  extruded = FALSE,
  elevation_scale = 1,
  wireframe = FALSE,
  line_width_units = "meters",
  line_width_scale = 1,
  line_width_min_pixels = 0,
  line_width_max_pixels = 9007199254740991,
  line_joint_rounded = FALSE,
  line_miter_limit = 4,
  get_fill_color = "#000000ff",
  get_line_color = "#000000ff",
  get_line_width = 1,
  get_elevation = 1000,
  material = TRUE,
  high_precision = FALSE,
  coverage = 1,
  center_hexagon = NULL,
  get_hexagon = hexagon,
  blending_mode = "normal",
  visibility_toggle = TRUE,
  tooltip = FALSE
)

Arguments

rdeck
<rdeck | rdeck_proxy> An rdeck map instance.

...
Additional parameters that will be forwarded to deck.gl javascript without validation nor processing. All dots must be named and will be camelCased when serialised. A warning is raised when dots are used, warning class rdeck_dots_nonempty.

id
<string> The layer's identifier must be unique for among all layers of the same type for a map. Defaults to uuid::UUIDgenerate(), but should be explicitly defined for updatable layers in a shiny application.

name
<string> Identifies the layer on tooltips and legends. It does not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.

data
<data.frame | sf | string> The layer's data. Data frames and sf objects will contain all columns that are referenced by the layer's accessors. Strings will be interpreted as a URL and data will be retrieved dynamically in the browser.

visible
<logical> Determines whether the layer is visible or not; also determines whether any legend elements for the layer will be displayed.

pickable
<logical> Determines if the layer responds to pointer / touch events.

opacity
<number> Determines the layer's opacity.

position_format
<"XY" | "XYZ"> Determines whether each coordinate has two (XY) or three (XYZ) elements.

color_format
<"RGB" | "RGBA"> Determines whether the alpha channel of the colours will be ignored by accessors, making all colours opaque.

auto_highlight
<logical> When TRUE, the current object hovered by the cursor is highlighted by highlight_color.

highlight_color
<accessor | scale | color> When auto_highlight and pickable are enabled, highlight_color determines the colour of the currently highlighted object. If a single colour value is supplied, that colour will be used to highlight all objects in the layer. Per-object highlighting is achieved with a colour scale, or a tidy-eval column of colours.

stroked
<logical> If TRUE, draw an outline around each object.

filled
<logical> If TRUE, draw the filled area of each point.

extruded
<logical> If TRUE, extrude objects along the z-axis; if FALSE, all objects will be flat.

elevation_scale
<number> The elevation multiplier.

wireframe
<logical> If TRUE and extruded == TRUE, draw a line wireframe of the object. The outline will have horizontal lines closing the top and bottom polygons and vertical lines for each vertex of the polygon.

line_width_units
<"pixels" | "meters"> The units of line width.

line_width_scale
<number> The line width multiplier.

line_width_min_pixels
<number> The minimum line width in pixels.

line_width_max_pixels
<number> The maximum line width in pixels.

line_joint_rounded
<logical>

line_miter_limit
number

get_fill_color
<accessor | scale | color> The fill colour of each object. Accepts a single colour value, a colour scale, or a tidy-eval column of colours.

get_line_color
<accessor | scale | color> The line colour of each object. Accepts a single colour value, a colour scale, or a tidy-eval column of colours.

get_line_width
<accessor | scale | number> The outline of the object in units specified by line_width_units. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers.

get_elevation
<accessor | scale | number> The elevation to extrude each object in the z-axis. Height units are in metres. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers.

material
<logical>

high_precision
<logical> If TRUE, draw each hexagon as a polygon using each hexagon's true geometry. If FALSE, draw each hexagon with the same shape as center_hexagon. High precision trades rendering performance for geometry precision. High precision is required (forcibly enabled) when:

Rendering hex resolutions in range c(0, 5)

Pentagons are present in the data

Multiple resolution hexes are present in the data

coverage
<number> Radius multiplier, between 0 - 1. The radius of each disk is calculated by coverage * radius.

center_hexagon
<string> If high_precision == FALSE, defines the H3 hexagon that will be used as the shape of all hexagons in the layer. Defaults to the hexagon in the centre of the viewport.

get_hexagon
<accessor> The H3 hex token for each H3 hexagon. Accepts a tidy-eval character column.

blending_mode
<"normal" | "additive" | "subtractive"> Sets the blending mode. Blending modes:

normal: Normal blending doesn't alter colours of overlapping objects.

additive: Additive blending adds colours of overlapping objects. Useful for highlighting dot density on dark maps.

subtractive: Subtractive blending darkens overlapping objects. Useful for highlighting dot density on light maps.

tooltip
<tooltip> Defines the columns (and their order) that will be displayed in the layer tooltip, if pickable == TRUE. Supports tidy-select if a data is a data.frame. sfc columns are always removed.

See also