MVT Layer

MVT Layer

Usage

add_mvt_layer(
  rdeck,
  ...,
  id = uuid::UUIDgenerate(),
  name = "MVTLayer",
  group_name = NULL,
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  wrap_longitude = FALSE,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_tile_data = NULL,
  extent = NULL,
  tile_size = 512,
  max_zoom = NULL,
  min_zoom = 0,
  max_cache_size = NULL,
  max_cache_byte_size = NULL,
  refinement_strategy = "best-available",
  z_range = NULL,
  max_requests = 6,
  zoom_offset = 0,
  unique_id_property = "",
  highlighted_feature_id = NULL,
  binary = TRUE,
  blending_mode = "normal",
  visibility_toggle = TRUE,
  tooltip = NULL,
  filled = TRUE,
  stroked = TRUE,
  line_width_max_pixels = 9007199254740991,
  line_width_min_pixels = 0,
  line_width_scale = 1,
  line_width_units = "meters",
  point_radius_max_pixels = 9007199254740991,
  point_radius_min_pixels = 0,
  point_radius_scale = 1,
  point_radius_units = "meters",
  point_antialiasing = TRUE,
  point_billboard = FALSE,
  get_fill_color = "#000000ff",
  get_line_color = "#000000ff",
  get_line_width = 1,
  get_point_radius = 1,
  icon_atlas = NULL,
  icon_mapping = list(),
  icon_size_max_pixels = 9007199254740991,
  icon_size_min_pixels = 0,
  icon_size_scale = 1,
  icon_size_units = "pixels",
  icon_alpha_cutoff = 0.05,
  icon_billboard = TRUE,
  get_icon = icon,
  get_icon_angle = 0,
  get_icon_color = "#000000ff",
  get_icon_pixel_offset = c(0, 0),
  get_icon_size = 1,
  text_size_max_pixels = 9007199254740991,
  text_size_min_pixels = 0,
  text_size_scale = 1,
  text_size_units = "pixels",
  text_background = FALSE,
  text_background_padding = c(0, 0, 0, 0),
  text_font_family = "Roboto, Helvetica, Arial, san-serif",
  text_font_weight = "normal",
  text_line_height = 1,
  text_max_width = -1,
  text_outline_color = "#000000ff",
  text_outline_width = 0,
  text_word_break = "break-word",
  text_billboard = TRUE,
  text_font_settings = list(),
  get_text = text,
  get_text_angle = 0,
  get_text_color = "#000000ff",
  get_text_pixel_offset = c(0, 0),
  get_text_size = 32,
  get_text_anchor = "middle",
  get_text_alignment_baseline = "center",
  get_text_background_color = "#ffffffff",
  get_text_border_color = "#000000ff",
  get_text_border_width = 0,
  line_joint_rounded = FALSE,
  line_cap_rounded = FALSE,
  line_miter_limit = 4,
  line_billboard = FALSE,
  extruded = FALSE,
  wireframe = FALSE,
  elevation_scale = 1,
  material = TRUE,
  get_elevation = 1000,
  point_type = "circle"
)

Arguments

rdeck

<rdeck | rdeck_proxy> An rdeck map instance.

...

Additional parameters that will be forwarded to deck.gl javascript without validation nor processing. All dots must be named and will be camelCased when serialised. A warning is raised when dots are used, warning class rdeck_dots_nonempty.

id

<string> The layer's identifier must be unique for among all layers of the same type for a map. Defaults to uuid::UUIDgenerate(), but should be explicitly defined for updatable layers in a shiny application.

name

<string> Identifies the layer on tooltips and legends. It does not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.

group_name

<string> Defines the group that this layer belongs to. Currently only effective on the layer selector, if visibility_toggle = TRUE.

data

<character> A character vector of mapbox vector tile url templates. Substrings "{x}", "{y}", "{z}" will be replaced with a tile's actual index on request.

If multiple url templates are supplied, each endpoint must return the same data for the same tile index.

visible

<boolean> Determines whether the layer is visible or not; also determines whether any legend elements for the layer will be displayed. When visble = NULL:

on update, use current state in the browser

on add, layer is visible

pickable

<boolean> Determines if the layer responds to pointer / touch events.

opacity

<number> Determines the layer's opacity.

wrap_longitude

<boolean> Normalises geometry longitudes.

position_format

<"XY" | "XYZ"> Determines whether each coordinate has two (XY) or three (XYZ) elements.

color_format

<"RGB" | "RGBA"> Determines whether the alpha channel of the colours will be ignored by accessors, making all colours opaque.

auto_highlight

<boolean> When TRUE, the current object hovered by the cursor is highlighted by highlight_color.

highlight_color

<accessor | scale | color> When auto_highlight and pickable are enabled, highlight_color determines the colour of the currently highlighted object. If a single colour value is supplied, that colour will be used to highlight all objects in the layer. Per-object highlighting is achieved with a colour scale, or a tidy-eval column of colours.

get_tile_data

<JS> retrieves the data of each tile. See deck.gl getTileData.

extent

<c(min_x, min_y, max_x, max_y)> Tiles in this bounding box will be rendered at min_zoom, when zoomed out below min_zoom.

tile_size

<number> A power of 2 that is the pixel dimensions of the tile.

max_zoom

<number> Tiles above this zoom level are not shown. Defaults to NULL.

min_zoom

<number> Tiles below this zoom level are not shown. Defaults to 0.

max_cache_size

<number> Maximum number of tiles that can be cached. Defaults to 5x the number of tiles in current viewport.

max_cache_byte_size

<number> Maximum memory used for caching tiles.

refinement_strategy

<"best-available" | "no-overlap" | "never"> How the tile layer refines visibility of tiles. Defaults to "best-available".

z_range

<c(min, max)> Array representing the range of heights in the tile.

max_requests

<number> Maximum number of concurrent HTTP requests across all specified tile provider domains. If a negative number is supplied no throttling occurs (HTTP/2 only).

zoom_offset

<int> The offset changes the zoom level at which the tiles are fetched.

unique_id_property

<string> Used for highlighting features across tiles. Features on separate tiles are deemed to be the same feature by the supplied property name. If no value is supplied, the feature id will be used.

highlighted_feature_id

<number | string> When provided, a feature with ID corresponding to the supplied value with be highlighted with highlight_color.

binary

<boolean> Improves rendering performance by removing the tile serialisation and deserialisation between worker and main thread.

Caveat: Polygons and multi-polygons with holes will be rendered as solid polygons.

blending_mode

<"normal" | "additive" | "subtractive"> Sets the blending mode. Blending modes:

normal: Normal blending doesn't alter colours of overlapping objects.

additive: Additive blending adds colours of overlapping objects. Useful for highlighting dot density on dark maps.

subtractive: Subtractive blending darkens overlapping objects. Useful for highlighting dot density on light maps.

visibility_toggle

<boolean> Whether this layer will appear in the layer selector.

tooltip

<tooltip> Defines the columns (and their order) that will be displayed in the layer tooltip, if pickable == TRUE. Supports tidy-select if a data is a data.frame. sfc columns are always removed.

filled

<boolean> If TRUE, draw the filled area of each point.

stroked

<boolean> If TRUE, draw an outline around each object.

line_width_max_pixels

<number> The maximum line width in pixels.

line_width_min_pixels

<number> The minimum line width in pixels.

line_width_scale

<number> The line width multiplier.

line_width_units

<"pixels" | "common" | "meters"> The units of line width.

point_radius_max_pixels

<number> The maximum radius in pixels.

point_radius_min_pixels

<number> The minimum radius in pixels.

point_radius_scale

<number> The radius multiplier for all points.

point_radius_units

<"pixels" | "common" | "meters"> The units of point radius.

get_fill_color

<accessor | scale | color> The fill colour of each object. Accepts a single colour value, a colour scale, or a tidy-eval column of colours.

get_line_color

<accessor | scale | color> The line colour of each object. Accepts a single colour value, a colour scale, or a tidy-eval column of colours.

get_line_width

<accessor | scale | number> The outline of the object in units specified by line_width_units. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers.

get_point_radius

<accessor | scale | number> The radius of each point, in units specified by radius_units. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers.

line_joint_rounded

<boolean>

line_miter_limit

number

extruded

<boolean> If TRUE, extrude objects along the z-axis; if FALSE, all objects will be flat.

wireframe

<boolean> If TRUE and extruded == TRUE, draw a line wireframe of the object. The outline will have horizontal lines closing the top and bottom polygons and vertical lines for each vertex of the polygon.

elevation_scale

<number> The elevation multiplier.

material

<boolean>

get_elevation

<accessor | scale | number> The elevation to extrude each object in the z-axis. Height units are in metres. Accepts a single numeric value, a numeric scale, or a tidy-eval column of numbers.

point_type

<"circle"|"icon"|"text"|combination> Determines how to render point and multipoint features. May be one of:

"circle"

"icon"

"text"

Or a combination, separated by "+", e.g. "circle+text".

See also