Tile Layer

Tile Layer

add_tile_layer(
  rdeck,
  ...,
  id = NULL,
  name = "TileLayer",
  data = NULL,
  visible = TRUE,
  pickable = FALSE,
  opacity = 1,
  position_format = "XYZ",
  color_format = "RGBA",
  auto_highlight = FALSE,
  highlight_color = "#00008080",
  get_tile_data = NULL,
  extent = NULL,
  tile_size = 512,
  max_zoom = NULL,
  min_zoom = 0,
  max_cache_size = NULL,
  max_cache_byte_size = NULL,
  refinement_strategy = "best-available",
  z_range = NULL,
  max_requests = 6,
  desaturate = 0,
  transparent_color = "#00000000",
  tint_color = "#ffffff",
  blending_mode = "normal",
  tooltip = FALSE
)

Arguments

rdeck
<rdeck | rdeck_proxy> An rdeck map instance.

...
Additional parameters that will be forwarded to deck.gl javascript without validation nor processing. All dots must be named and will be camelCased when serialised. A warning is raised when dots are used, warning class rdeck_dots_nonempty.

id
<string> The layer's identifier must be unique for among all layers of the same type for a map. Defaults to uuid::UUIDgenerate(), but should be explicitly defined for updatable layers in a shiny application.

name
<string> Identifies the layer on tooltips and legends. It does not need to be unique, but should be brief. Defaults to the deck.gl class name for the layer.

data
<character> A character vector of raster tile url templates. Substrings "{x}", "{y}", "{z}" will be replaced with a tile's actual index on request.

If multiple url templates are supplied, each endpoint must return the same data for the same tile index.

visible
<logical> Determines whether the layer is visible or not; also determines whether any legend elements for the layer will be displayed.

pickable
<logical> Determines if the layer responds to pointer / touch events.

opacity
<number> Determines the layer's opacity.

position_format
<"XY" | "XYZ"> Determines whether each coordinate has two (XY) or three (XYZ) elements.

color_format
<"RGB" | "RGBA"> Determines whether the alpha channel of the colours will be ignored by accessors, making all colours opaque.

auto_highlight
<logical> When TRUE, the current object hovered by the cursor is highlighted by highlight_color.

highlight_color
<accessor | scale | color> When auto_highlight and pickable are enabled, highlight_color determines the colour of the currently highlighted object. If a single colour value is supplied, that colour will be used to highlight all objects in the layer. Per-object highlighting is achieved with a colour scale, or a tidy-eval column of colours.

get_tile_data
<JS> retrieves the data of each tile. See deck.gl getTileData.

extent
<c(min_x, min_y, max_x, max_y)> Tiles in this bounding box will be rendered at min_zoom, when zoomed out below min_zoom.

tile_size
<number> A power of 2 that is the pixel dimensions of the tile.

max_zoom
<number> Tiles above this zoom level are not shown. Defaults to NULL.

min_zoom
<number> Tiles below this zoom level are not shown. Defaults to 0.

max_cache_size
<number> Maximum number of tiles that can be cached. Defaults to 5x the number of tiles in current viewport.

max_cache_byte_size
<number> Maximum memory used for caching tiles.

refinement_strategy
<"best-available" | "no-overlap" | "never"> How the tile layer refines visibility of tiles. Defaults to "best-available".

z_range
<c(min, max)> Array representing the range of heights in the tile.

max_requests
<number> Maximum number of concurrent HTTP requests across all specified tile provider domains. If a negative number is supplied no throttling occurs (HTTP/2 only).

desaturate
<number> The desaturation of the bitmap. Between 0 and 1, being the original colour, 1 being greyscale.

transparent_color
<color> The colour to use for transparent pixels.

tint_color
<color> The colour to tint the bitmap by. Alpha channel is ignored if supplied.

blending_mode
<"normal" | "additive" | "subtractive"> Sets the blending mode. Blending modes:

normal: Normal blending doesn't alter colours of overlapping objects.

additive: Additive blending adds colours of overlapping objects. Useful for highlighting dot density on dark maps.

subtractive: Subtractive blending darkens overlapping objects. Useful for highlighting dot density on light maps.

tooltip
<tooltip> Defines the columns (and their order) that will be displayed in the layer tooltip, if pickable == TRUE. Supports tidy-select if a data is a data.frame. sfc columns are always removed.

See also