Prop Type: Accessor

Accessors map data columns to visual representations, primarily colours: fill, line, highlight and sizes: radius, elevation, width, height.

Accessors support tidy-eval expressions. String literals referencing a column aren't supported (this will be interpreted as a literal string), however you may pass a bare name (my_col) or an unquoted variable (!!my_var or {{my_var}}).

On the client, an accessor is translated to a javascript function that retrieves a value from data for each object rendered. Constant columns will work fine, but aren't recommended as they are less performant on render and bloat file size; prefer a constant value for these cases.

See accessors for details.

Accessors & Scales

Accessors may be used everywhere that scales can and can perform almost the same function (i.e. mapping data to a vector of numbers or colours), but there are downsides:

larger file size

no legend support

Prefer scales over accessors where both are supported, and use accessors where scales aren't supported (e.g. get_path) or where scaling functions aren't appropriate.