Scale category

Creates a categorical scale. Categorical scales map input values defined in the set of levels to colours (or values). Input values not in the set of levels are assigned unmapped_color (or unmapped_value).

scale_color_category(
  col,
  palette,
  unmapped_color = "#000000",
  levels = NULL,
  unmapped_tick = NULL,
  tick_format = NULL,
  legend = TRUE
)

scale_category(
  col,
  range = 0:1,
  unmapped_value = 0,
  levels = NULL,
  legend = TRUE
)

Arguments

col
<name | string> The name of the column containing data to be scaled. Can be either a named column (non-standard evaluation), or an expression evaluating a string. Supports tidy-eval.

palette
<color> The colour palette of the colour scale. Must be a vector of rgba hex strings of length >= 2.

unmapped_color
<color> The colour representing unmapped levels.

levels
<factor | character | logical> The category levels. If NULL, will be populated from input data. The order of the levels is determined by levels() for factors & unique() otherwise. Length of levels must equal palette or range, such that each category level is assigned a colour or range value.

unmapped_tick
<string> The tick label of the unmapped category. If not NULL and legend == TRUE, the unmapped category will appear at the bottom of the legend.

tick_format
<function> A function taking a vector of ticks returning formatted ticks.

legend
<logical> Indicate whether the legend should be displayed for this scale.

range
<numeric> The output range of the numeric scale. Must be be a numeric vector of length >= 2.

unmapped_value
<number> The value representing unmapped levels.

See also